Replaced with a type float to decimal (18,8) in the column strikeprice

+++ ICE_Import/ICE_Import/StoredProcs/cqgdb.test_SPO.sql
@@ -3,7 +3,7 @@ CREATE PROCEDURE [cqgdb].[test_SPO]
     @optionmonth CHAR (1), 
     @optionmonthint INT, 
     @optionyear INT, 
-    @strikeprice FLOAT (53), 
+    @strikeprice DECIMAL (18,8), 
     @callorput CHAR (1), 
     @idinstrument BIGINT, 
     @expirationdate DATE,
